+++ services/music-service/src/MusicService.Api/Program.cs
@@ -1,4 +1,5 @@
 using MusicService.Api.Middleware;
+using MusicService.Application;
 using MusicService.Infrastructure;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -8,6 +9,7 @@ builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
 builder.Services.AddHttpContextAccessor();
 builder.Services.AddInfrastructure(builder.Configuration);
+builder.Services.AddApplication();
 
 var app = builder.Build();